Implement functionality to deal with large files
Previous versions of the code needed to read the entire VCF file and
hold on to many values in there (in particular, for each individual, a
full set of the variation data was kept on memory). For large files,
this would result in large amounts of RAM being consumed. The new
version updates the state of a comparer and keeps only a small amount of
data for each individual and for each pair of individuals (in effect,
only the "current" distance being calculated)