Move some variant pre-processing to the correct method
The comparison was done on a list of alleles, but that list could have
been automatically converted to a set and processed once for each
individual, instead of in all comparison. This reduced the amount of
time spent on the actual update_comparison method to about 28%, while
increasing the time of pre-processing the variants a negligible amount.